Continuation of the modif, new folder organization
